Remove warning about status of AllowUnexported
At this point, we will not be removing the AllowUnexported option
as there are several use-cases that legitimately benefit from it.
Also, the current documentation does a good job of explaining
what the proper use of it is. Thus, remove the warning as it is
unhelpful and confusing to users.

The other (unlikely) alternative I can see in the future is to
make AllowUnexported the default behavior for all comparisons
effectively making AllowUnexported a noop. That would be a
backwards compatible change.