Add Last helper method
It is very common for FilterPath functions to operate on the last step
in a path. However, it is a hindrance to continually write:
    if len(p) == 0 {
        return false
    }
    return p[len(p)-1].Type() == ...

Instead, add a helper method Last that avoids the empty path check
and the indexing to the last element.
This allows users to simply write:
    return p.Last().Type() == ...

This is more readable and writeable.