Expand dynamic checks of options (#20)
Add checks for mutations of the input value and non-deterministic
transformer outputs.

A prior approach (not shown in this commit) attempted to detect mutations of
the input by hashing the inputs before and after calling the option functions.
However, we deemed this approach too restrictive since there are legitimate
cases where some types lazily update their values upon first access
(for example, accessing extensions on proto2 messages). This would cause the
hashes to mismatch and provide a false-positive panic.

In this approach, we accept the reality that lazy evaluators exist, but make
the assumption that any lazy evaluators are protected by a mutex of some sort.
Thus, we simply run the option function twice in parallel, and let the race
detector detect mutations on the input. We intentionally run the options
in a function named detectRaces to make it obvious from the stack trace
what's going wrong.

For the detection of deterministic output of Transformers, we call a special
version of s.compareAny so that the outputs may be compared according to the
semantics of what the user provided. The specialized wrapper s.statelessCompare
preserves the comparison state such as the current result and the reporter.