Check for probable infinite recursive cycles
After some hard-coded limit, check the path for recursive transformers.
Note that detecting an infinite cycle of recursive transformers
would be equivalent to solving the halting problem,
so we could misdiagnose a valid recursion as infinite.

For this reason, the check only triggers after some minimum stack size
so that correctness is not compromised for nearly all situations.
However, the threshold is low enough to trigger before the Go runtime
panics with a stack overflow (which is not recoverable).

Example panic message:
<<<
panic: recursive set of Transformers detected:
	Transformer(T1, main.main.func1): complex64 => complex128
	Transformer(T2, main.main.func2): complex128 => [2]float64
	Transformer(T3, main.main.func3): float64 => complex64
consider using cmpopts.AcyclicTransformer
>>>

Updates #77