Export the Reporter API
The Reporter option allows users to hook in their own custom reporters
to programatically interpret the diff structure.

The Reporter API uses a push/pop mechanism, which is strictly more powerful
than an API that only calls report on leaf nodes.
The Reporter.Report method takes in a Result type to provide flexibility in
what properties can be reported in the future since new properties can be added,
but new methods cannot be easily added to the reporter interface.