cmp/internal/value: fix handling of negative zero for floats
Fix IsZero to properly report false for IsZero(-0.0) since
we define IsZero as whether it is equal to the zero memory value.
Add note to isLess that we don't need to handle -0.0 since
we can't possibly have both keys present in the same map.
Use sort.SliceStable in SortedKeys for deterministic output since
it is possible to have both -0.0 and +0.0 from two different maps.