Allow batched diffing of slices with a custom comparer
For correctness, cmp checks applicability of the options for every
element in a slice. For large []byte, this is a significant performance
detriment. The workaround is to specify Comparer(bytes.Equal).
However, we would still like to have the batched diffing if the
slices are different. Specialize for this situation.