Introduce deliberate instability to difference output
The reporter output is documented as unstable.
The API for custom reporters also specifies that the diffing of
slices is unstable.

Introduce deliberate instability to the diffing algorithm so that
we have the flexibility to improve it in the future.
The current algorithm optimizes for speed, rather than optimality,
so there is much room for improvement.