Special-case for top-level triple-quoted string literals
For top-level triple-quoted string literals, the surrounding parenthesis
is an unnecessary amount of verbosity and indentation.
These parenthesis are always needed for any situation where the triple-quoted
string literal does not appear at the top-level to ensure that it nests well
within some other composite structure.

Thus, we detect triple-quoted string literals at the top-level
and format them verbatim. The textList.formatExpandedTo method
assumes that it is usually called within the context of a composite structure,
so we perform some string mangling to correct the leading and trailing bytes.

Fixes #301