Do not inhibit the dead code elimination.
Avoid calls to value.TypeString() that do not request fully qualified
type names. The implementation of TypeString() depends on getting
methods by index which disables the DCE in the compiler.

k8s.io/kube-openapi and k8s.io/apiserver happen to depend on go-cmp
outside tests so it is important that go-cmp does not inhibit the DCE
inadvertently.